preco

+++ Estrutura_Controle/Estrutura_Sequecial/soma_quatro_inteiros.py
@@ -2,10 +2,10 @@
 ## AUTOR: PROF. GUILHERME CARDOSO AGOSTINETTI
 
 # ENTRADA DE DADOS
-N1 = float(input("N1: "))
-N2 = float(input("N2: "))
-N3 = float(input("N3: "))
-N4 = float(input("N4: "))
+N1 = int(input("N1: "))
+N2 = int(input("N2: "))
+N3 = int(input("N3: "))
+N4 = int(input("N4: "))
 
 # PROCESSAMENTO 
 SOMA = N1+N2+N3+N4
